move ANativeWindow and friends under system/core
This patch moves the definition of ANativeWindow from frameworks/base into
system/core.  We move ANativeWindow for two reasons.

Firstly, the native-window interface may need to be used by lower-level code
than the framework.  Such code (such as the camera HAL) is not allowed to use
definitions from frameworks/base, since the latter is stacked on top of the
former.  Thus moving (the interface definition) ANativeWindow under system/core
lowers the level at which we can work with windows.

Secondly, this enables an implementation of ANativeWindow that lies outside of
frameworks, enabling quick initial bringup of graphics.

Change-Id: I851b9bb2834bcec73255b5af82f66411e54bb1a5
Signed-off-by: Iliyan Malchev <malchev@google.com>