Add HAL_PIXEL_FORMAT_BLOB
This format describes buffers that are not structured like normal
image data, and can only be meaningfully understood by the endpoints.

As an example, this format is used for transporting JPEG-compressed
image buffers from the Camera HAL to the framework/apps.

Buffers of this format must have a height of 1 and a width equal to
their size in bytes. Otherwise their interpretation is up to the two
endpoints.

Bug: 6243944
Change-Id: Ia81015694cbf3a89e29bfbf3624c3ea2ef4f66fd