crash_sender: add 30sec holdoff for fresh crashes
As a rule, the .meta file of a crash report should only be
created after the corresponding crash data files have been
written to avoid the crash report to be sent prematurely without
data files.  Merely as an additional safeguard, this changeset
prevents a crash report to be sent within 30 seconds of writing
the .meta file.  This functionality integrates with the existing
sleep due to SECONDS_SEND_SPREAD.

BUG=chromium:338977
TEST=`cbuildbot x86-generic-full` passes

Change-Id: Id3651eb697527ca5b976274933b0ca2ee09870e2
Reviewed-on: https://chromium-review.googlesource.com/200061
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>