init.trace.rc: fix race condition with class_start core
The systrace permissions from init.trace.rc and the "class_start core"
which launches surfaceflinger are both in an "on boot" section. However,
the init.trace.rc commands are parsed after all commands in init.rc.
This means that "class_start core" is executed before the chmod command
which allows processes to write to trace_marker. If any services
execute their first trace command before the chmod occurs, then that
service won't be able to write traces until the service is restarted.

To fix this, run all of the init.trace.rc commands in the "early-boot"
section to ensure they are completed first.

Bug: 17612265
Change-Id: Ibf544762173d5ba98272c66ef485d8eab7d70bf3