Created service 'bugreportplus'
This service is an enhanced version of bugreport that provides a better
user interface (like displaying progress and allowing user to enter
details).

It will be  typically triggered by the 'Take Bug Report' UI, which will
now offer the option for the traditional or enhanced options (services
'bugreport' and 'bugreportplus' respectively).

BUG: 26034608
Change-Id: I39ea92c3e329a801b51f60a558c73faaf890c068